requirements.txt
aiogram
celery
redis
python-dotenv